调整项目为可调用

+++ admin/src/main/java/com/igeek/admin/model/vo/UsersVo.java
@@ -11,7 +11,7 @@ import java.io.Serializable;
 @Data
 public class UsersVo implements Serializable {
 
-    Long code;
+    Integer code;
     String message;
     String accout;
     String token;
